Title: 트리에서 공 떨어뜨리기, Time: 126ms, Memory: 1MB, Status: Passed - CodeTree

+++ 240119/트리에서 공 떨어뜨리기/drop-a-ball-in-a-tree.cpp
@@ -11,7 +11,7 @@ int main() {
 	for (int i = 1; i <= n; i++) {
 		cin >> lnode[i] >> rnode[i];
 	}
-	int k;
+	long long k;
 	cin >> k;
 	int cur = 1;
 	while (1) {
